% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfList-class.R
\docType{class}
\name{cfList-class}
\alias{cfList-class}
\title{An S4 class cfList}
\description{
Class \code{cfList} for most cytofast functions
}
\section{Slots}{

\describe{
\item{\code{samples}}{data frame containing information on each sample}

\item{\code{expr}}{data frame containing the marker expression}

\item{\code{counts}}{data frame containing (normalized) cell counts per cluster and per sample}

\item{\code{results}}{data frame containing results}
}}

