% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfList-constructor.R
\name{cfList}
\alias{cfList}
\title{Constructor for a cfList}
\usage{
cfList(samples = NULL, expr = NULL, counts = NULL)
}
\arguments{
\item{samples}{data frame containing information on each sample}

\item{expr}{data frame containing the marker expression}

\item{counts}{data frame containing (normalized) cell counts per cluster and per sample}
}
\value{
The function returns an object of class \code{cfList}.
}
\description{
Creates a \code{cfList} S3 object.
}
\examples{
# Create empty cfList
cfData <- cfList()


}
\keyword{cf}
\keyword{cytofast,}
\keyword{list,}
