% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenateFCS.R
\name{concatenateFCS}
\alias{concatenateFCS}
\title{Merge .fcs files in selected folder}
\usage{
concatenateFCS(fcsDir, filename, path = NULL)
}
\arguments{
\item{fcsDir}{Directory with .fcs files to be combined}

\item{filename}{File name to create on disk}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename} are
combined to create the fully qualifed file name. Defaults to \code{fcsDir}}
}
\value{
A character vector of the file name.
}
\description{
Function to merge .fcs files in a given folder.
}
\examples{
concatenateFCS(fcsDir = system.file("extdata", package="cytofast"), 
               filename = merged/cytofastData_merged.fcs)

}
\keyword{.fcs}
\keyword{files}
\keyword{merge}
