% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoBoxplots.R
\name{cytoBoxplots}
\alias{cytoBoxplots}
\title{Draw boxplots for cfList}
\usage{
cytoBoxplots(cfList, group, stat)
}
\arguments{
\item{cfList}{a \code{\link[cytofast]{cfList}} object. It should contain at least data in the 'counts' slot.}

\item{group}{one of:
\itemize{
\item a character referring to a column name in the \code{samples} slot of the \code{cfList}.
\item a factor indicating the grouping (x-axis) for the boxplots.
}}

\item{stat}{ignore, will be implemented soon.}
}
\description{
Draw boxplots for all given clusters. Values are based on the \code{counts} slot in the
\code{cfList}.
}
\keyword{boxplot,}
