% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoExprScale.R
\name{cytoExprScale}
\alias{cytoExprScale}
\title{Scale the expression values of a cfList object}
\usage{
cytoExprScale(cfList, scale = "clusterMax", expr.limits = c(0, 6))
}
\arguments{
\item{cfList}{a cfList object.}

\item{expr.limits}{is a vector of to numerical values, the minimum and the
maximum of the expression values.}

\item{scale:}{\itemize{
\item clusterMax: for each marker, there is a cluster that has an
expression value that is at maximum. So the expression values are scaled
such the maximal value of a cluster is top limit. The scaling does not
change the zero: it is just a multiplicative coefficient.
}}
}
\value{
The function returns an object of class
\code{\link[cytofast]{cfList}}.
}
\description{
This function scales the expression values.
}
\examples{

}
\keyword{FCS,}
\keyword{expression}
