% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoHeatmaps.R
\name{cytoHeatmaps}
\alias{cytoHeatmaps}
\title{Draw heatmaps for cfList}
\usage{
cytoHeatmaps(cfList, group, legend = FALSE)
}
\arguments{
\item{cfList}{a \code{\link[cytofast]{cfList}} object.}

\item{group}{one of:
\itemize{
\item a character vector referring to a column name in the \code{samples} slot of the \code{cfList}.
\item a factor indicating the grouping (x-axis) for the boxplots.
}}

\item{legend}{logical, whether a legend should be added}
}
\description{
Function to draw two heatmaps. They visualize the median intensity of the markers for the
created clusters. The ordering of the clusters is based on the default
hierarchical cluster analysis \code{\link[stats]{hclust}}. Note that hclust takes the data after
the median intensity is calculated per cluster, thus placing the most similar clusters
next to each other.
}
\keyword{FCS}
\keyword{heatmap,}
\keyword{markers,}
