% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytottest.R
\name{cytottest}
\alias{cytottest}
\title{preform t-test on cfList}
\usage{
cytottest(cfList, group, adjustMethod, ...)
}
\arguments{
\item{cfList}{a \code{\link[cytofast]{cfList}} object. It should contain at least data in the 'counts' slot.}

\item{group}{one of:
\itemize{
\item a character vector referring to a column name in the \code{samples} slot of the \code{cfList}.
\item a factor indicating the grouping for the t.test.
}}

\item{adjustMethod}{character, correction method used in \code{\link[stats]{p.adjust}}, choose from
\code{c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}. If omitted, no correction takes place.}

\item{...}{further arguments passed on to \code{\link[stats]{t.test}}.}
}
\description{
Performs a separate t-test on each cluster within a cfList. The output is added and can also be used
by other functions
}
