% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_IO-data.R
\docType{data}
\name{material_io}
\alias{material_io}
\title{Color palette from material.io}
\format{
A data frame with 14 observations and 19 variables:
\describe{
\item{red}{shades of red}
\item{pink}{shades of pink}
\item{purple}{shades of purple}
\item{deep.purple}{shades of deep purple}
\item{indigo}{shades of indigo}
\item{blue}{shades of blue}
\item{light.blue}{shades of light blue}
\item{cyan}{shades of cyan}
\item{teal}{shades of teal}
\item{green}{shades of green}
\item{light.green}{shades of light green}
\item{lime}{shades of lime}
\item{yellow}{shades of yellow}
\item{amber}{shades of amber}
\item{orange}{shades of orange}
\item{deep.orange}{shades of deep orange}
\item{brown}{shades of brown}
\item{gray}{shades of gray}
\item{blue.gray}{shades of blue.gray}
}
}
\source{
\url{https://material.io/resources/color/}
}
\usage{
material_io
}
\description{
A dataset containing hexcodes for several colors
}
\keyword{datasets}
