% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfList-constructor.R
\name{print.cfList}
\alias{print.cfList}
\title{Print method for cfList}
\usage{
\method{print}{cfList}(x, ...)
}
\arguments{
\item{x}{a cfList object}

\item{...}{further arguments passed to or from other methods}
}
\value{
None, but prints the str() of an cfList.
}
\description{
Print method for cfList
}
\examples{
# Create empty cfList
cfData <- cfList()

# Print will return the str() of an cfList
cfData

}
