% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCytofkitFCS.R
\name{readCytofkitFCS}
\alias{readCytofkitFCS}
\title{Read .fcs files created by Cytofkit and return a cytofast object}
\usage{
readCytofkitFCS(dir = NULL, colNames = c("description", "names"),
  clusterID = "FlowSOM_clusterIDs", clean.up = TRUE)
}
\arguments{
\item{dir}{directory containing the .FCS files created by Cytofkit}

\item{colNames}{character string that specifies which label should be used as
the column names. This could be the name of the markers 'names' or the
description for the markers 'description'.}

\item{clusterID}{character string that specifies which label should be used
to retrieve the cluster identifiers. Typically, this should be
"FlowSOM_clusterIDs", "Rphenograph_clusterIDs" or "ClusterX_clusterIDs"}

\item{cleanup}{logical that specifies to remove extra channels added by
Cytofkit}
}
\value{
The function returns an object of class
\code{\link[cytofast]{cfList}}.
}
\description{
This function reads and combines .FCS files created specifically
by Cytofokit.
}
\note{
To successfully read the .FCS files created by cytofkit make sure that
there are no double channels. It is possible to remove the double tag in
software like Flowjo.

As for the \code{colNames} parameter, both 'name' and 'description' will always
work, it is up to the user to decide which one is most preferable. "" or
"<NA>" descriptions are replaced by the corresponding name.

This function is a wrapper around \code{\link[flowCore]{read.FCS}}. For
more flexibility see their help page. Parameters currently set are:
transformation=FALSE, truncate_max_range=FALSE, min.limit = NULL

NB: No compensation nor transformation is applied to the data.
}
\examples{
dirFCS <- system.file("extdata", package="cytofast")
cfData <- readCytosploreFCS(dir = dirFCS, colNames = "description")

}
\keyword{FCS,}
\keyword{cytofkit}
\keyword{data,}
\keyword{read,}
