% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfList-methods.R
\docType{methods}
\name{show,cfList-method}
\alias{show,cfList-method}
\title{Show cfList object}
\usage{
\S4method{show}{cfList}(object)
}
\arguments{
\item{object}{a cfList object}
}
\value{
Prints a light version of a \code{cfList} to the console
}
\description{
Show cfList object
}
