% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfList-constructor.R
\docType{methods}
\name{show,cfList-method}
\alias{show,cfList-method}
\title{show method for cfList}
\usage{
\S4method{show}{cfList}(object)
}
\arguments{
\item{object}{a cfList object}
}
\value{
None, but prints the str() of an cfList.
}
\description{
show method for cfList
}
\examples{
# Create empty cfList
cfData <- cfList()

# show will return the str() of an cfList
cfData

}
