% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spitzer-data.R
\docType{data}
\name{spitzer}
\alias{spitzer}
\title{Spitzer meta file}
\format{A data frame with 26 observations and 3 variables:
\describe{
\item{CSPLR_ST}{sample tag, corresponds with .fcs files created by Cytosplore}
\item{group}{grouping variable, effected vs. ineffected, and day 3 vs. day 8}
\item{sample_name}{original sampleID}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5312823/}
}
\usage{
spitzer
}
\description{
A dataset containing information on the samples from the Spitzer study
}
\references{
\insertRef{Spitzer2017}{cytofast}
}
\keyword{datasets}
