% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfList-methods.R
\docType{methods}
\name{[,cfList,ANY-method}
\alias{[,cfList,ANY-method}
\title{Extract parts of cfList}
\usage{
\S4method{[}{cfList,ANY}(x, i, j)
}
\arguments{
\item{x}{a cfList object from which to extract from}

\item{i}{index specifying which samples to extract. The index is either
a \code{numeric} or \code{character} vector.}

\item{j}{index specifying which clusters to extract. The index is either
a \code{numeric} or \code{character} vector.}
}
\value{
Returns the specified clusters or samples from a \code{cfList}.
}
\description{
Extract parts of cfList
}
