% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCounts.R
\name{cellCounts}
\alias{cellCounts}
\title{Extract cell counts from cfList}
\usage{
cellCounts(cfList, frequency = FALSE, scale = FALSE)
}
\arguments{
\item{cfList}{a cfList object. It should contain at least data in the 'expr' slot.}

\item{frequency}{one of
\itemize{
\item a logical value. if \code{FALSE}, the abundance of the cell counts are used. If
\code{TRUE}, the frequency of the total amount of given cells is used.
\item a numeric vector of same length as amount of samples
}}

\item{scale}{a logical value. Do the cell frequencies need to be centered and scaled? The
default \code{\link[base]{scale}} function is called.}
}
\value{
Returns the given cfList with a 'counts' slot.
}
\description{
A function to add the frequency (or abundance) of cell clusters per sample to a
\code{cfList}.
}
\examples{
# Read Data
dirFCS <- system.file("extdata", package="cytofast")
cfData <- readCytosploreFCS(dir = dirFCS, colNames = "description")

# Add cell counts to cfList
cfData <- cellCounts(cfData)


}
\keyword{cells}
\keyword{counts,}
\keyword{frequency,}
