% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfList-class.R, R/cfList-accessors.R,
%   R/cfList-methods.R
\docType{class}
\name{cfList-class}
\alias{cfList-class}
\alias{cfList}
\alias{samples}
\alias{samples,cfList-method}
\alias{expr}
\alias{expr,cfList-method}
\alias{counts}
\alias{counts,cfList-method}
\alias{results}
\alias{results,cfList-method}
\alias{show,cfList-method}
\title{Class cfList}
\description{
This is an S4 class that contains all the data and results
for the cytofast workflow. Most functions in this package will both
use and return a \code{cfList}.
}
\details{
This class is used througout the cytofast workflow, most functions in this
package will both use and return a \code{cfList}. FCS files created by
Cytosplore can be easily read in with \code{\link{readCytosploreFCS}}) and will
be returned as a cfList. It is also possible to manually create instances of this
class if another clustering method is preferred. Below an example is
shown for \code{\link[FlowSOM]{FlowSOM}}.
}
\section{Slots}{

\describe{
\item{\code{samples}}{data frame containing all meta information on each sample}

\item{\code{expr}}{data frame containing the marker expression}

\item{\code{counts}}{data frame containing (standardized) cell counts per cluster and per sample}

\item{\code{results}}{list containing any results}
}}

\section{Methods}{

\describe{
\item{\code{samples}}{a method for obtaining the samples data}
\item{\code{expr}}{a method for obtaining the marker expression}
\item{\code{counts}}{a method for obtaining the cell counts}
\item{\code{results}}{a method for obtaining any results from performed tests}
}
}

\examples{
### manually create instance of cfList class
library(FlowSOM)

## Cluster with FlowSOM
dirFCS <- system.file("extdata", package="cytofast")
fSOM <- FlowSOM(input = dirFCS, 
                transform = FALSE,
                scale = FALSE,
                colsToUse = c(9:11, 15:52),
                nClus = 10, # Note that this is an ambiguous choice
                seed = 123)

## expr slot
# retrieve clusters
clusterID <- as.factor(fSOM$FlowSOM$map$mapping[,1])
levels(clusterID) <- fSOM$metaclustering

# retrieve samples (As example, we assume each FCS file is its own sample)
sampleID <- lapply(fSOM$FlowSOM$metaData, function(x){rep(x[1], each = length(x[1]:x[2]))})
attr(sampleID, 'names') <- NULL
sampleID <- as.factor(unlist(sampleID)) 
levels(sampleID) <- paste("ID", 1:10, sep="_")

exprD <- data.frame(clusterID, 
                    sampleID, 
                    fSOM$FlowSOM$data[, c(9:11, 15:52)])

## samples slot
samplesD <- data.frame(sampleID = levels(sampleID), 
                       group = rep(c("group1", "group1"), each=5)) 

## create cfList
cfList(samples = samplesD,
       expr = exprD)

}
