% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoBoxplots.R
\name{cytoBoxplots}
\alias{cytoBoxplots}
\title{Draw boxplots for cfList}
\usage{
cytoBoxplots(cfList, group, stat)
}
\arguments{
\item{cfList}{a cfList object. It should contain at least data in the 'counts' slot.}

\item{group}{one of:
\itemize{
\item a character referring to a column name in the \code{samples} slot of the \code{cfList}.
\item a factor indicating the grouping (x-axis) for the boxplots.
}}

\item{stat}{ignore, will be implemented soon.}
}
\value{
None
}
\description{
Draw boxplots for all given clusters. Values are based on the \code{counts} slot in the
\code{cfList}.
}
\examples{
# Read Data
dirFCS <- system.file("extdata", package="cytofast")
cfData <- readCytosploreFCS(dir = dirFCS, colNames = "description")

# Add cell counts to cfList and add meta data
cfData <- cellCounts(cfData, frequency = TRUE, scale = TRUE)
meta <- spitzer[match(row.names(cfData@samples), spitzer$CSPLR_ST),]
cfData@samples <- cbind(cfData@samples, meta)

# Remove unnecessary markers
cfData@expr <- cfData@expr[,-c(3:10, 13:16, 55:59, 61:63)]

# Draw boxplots
cytoBoxplots(cfData, group="group")


}
\keyword{boxplot,}
