% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytottest.R
\name{cytottest}
\alias{cytottest}
\title{preform t-test on cfList}
\usage{
cytottest(cfList, group, adjustMethod, ...)
}
\arguments{
\item{cfList}{a \code{\link[cytofast]{cfList}} object. It should contain at least data in the 'counts' slot.}

\item{group}{one of:
\itemize{
\item a character vector referring to a column name in the \code{samples} slot of the \code{cfList}.
\item a factor indicating the grouping for the t.test.
}}

\item{adjustMethod}{character, correction method used in \code{\link[stats]{p.adjust}}, choose from
\code{c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}. If omitted, no correction takes place.}

\item{...}{further arguments passed on to \code{\link[stats]{t.test}}.}
}
\value{
Returns a cfList with a 'results' slot.
}
\description{
Performs a separate t-test on each cluster within a cfList. The output is added and can also be used
by other functions
}
\examples{
# Read Data
dirFCS <- system.file("extdata", package="cytofast")
cfData <- readCytosploreFCS(dir = dirFCS, colNames = "description")

# relabeling of clusterID
levels(cfData$expr$clusterID) <- gsub("[^0-9]", "", levels(cfData$expr$clusterID))  

# Add cell counts to cfList and add meta data
cfData <- cellCounts(cfData, frequency = TRUE, scale = TRUE)
meta <- spitzer[match(row.names(cfData$samples), spitzer$CSPLR_ST),]
cfData$samples <- cbind(cfData$samples, meta)

# Run t-test
cfData$samples$effect <- gsub("_D\\\\d", "", spitzer$group)
cfData <- cytottest(cfData, group  = "effect", adjustMethod = "bonferroni")
cfData$results

}
