% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawTree.R
\name{drawTree}
\alias{drawTree}
\title{Draw a dendogram with ggplot}
\usage{
drawTree(hclust)
}
\arguments{
\item{hclust}{an object of class \code{\link[stats]{hclust}}.}
}
\description{
Function to draw a dendrogram with ggplot, the tree can be easily alligned to other graphs (e.g. heatmaps).
}
\keyword{dendrogram}
\keyword{heatmap}
\keyword{tree}
