% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfList-constructor.R
\name{print.cfList}
\alias{print.cfList}
\title{Print method for cfList}
\usage{
\method{print}{cfList}(x, ...)
}
\arguments{
\item{x}{a cfList object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print method for cfList
}
