% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCytosploreFCS.R
\name{readCytosploreFCS}
\alias{readCytosploreFCS}
\title{Read .fcs files created by cytosplore}
\usage{
readCytosploreFCS(dir = NULL, colNames = c("names", "description"))
}
\arguments{
\item{dir}{directory containing the .FCS files created by Cytosplore}

\item{colNames}{character string that specifies which label should be used as the column names. This
could be the name of the markers 'names' or the description for the markers 'description'.}
}
\value{
The function returns an object of class \code{\link[base]{data.frame}}. It includes both
\code{clusterID} and \code{sampleID} as variables.
}
\description{
This function reads and combines .FCS files created specifcally by Cytosplore.
}
\note{
This function is a wrapper around \code{\link[flowCore]{read.FCS}}. For more
flexibility see their help page.
}
\keyword{FCS,}
\keyword{cytosplore}
\keyword{data,}
\keyword{read,}
